+++ src/components/SimpleChart/types.ts
@@ -1,0 +1,4 @@
+export interface StockDataProp {
+  date: string;
+  close: number;
+}
